修复某些id是int的问题、 声音选择seekbar、 修复首页没沉浸问题

+++ common_bokang/src/main/java/com/yijia/common_yijia/main/index/YjReFreshHandler.java
@@ -120,7 +120,7 @@ public class YjReFreshHandler extends RefreshHandler {
         final int currentCount = BEAN.getCurrentCount();
         final int total = BEAN.getTotal();
         final int index = BEAN.getPageIndex();
-        final int BeginCircleId = BEAN.getBeginCircleId();
+        final long BeginCircleId = BEAN.getBeginCircleId();
         if(index<=1){
             return;
         }
